requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
keras==2.13.1
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.22

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
bokeh==3.2.1

# Jupyter and Development
jupyter==1.0.0
jupyterlab==4.0.3
ipykernel==6.24.0

# Data Processing
dask==2023.7.1
polars==0.18.8

# Model Deployment and API
fastapi==0.101.0
uvicorn==0.23.1
pydantic==2.1.1

# MLOps and Experiment Tracking
mlflow==2.5.0
wandb==0.15.7
dvc==3.12.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
hydra-core==1.3.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3

# Security
bandit==1.7.5

# Utilities
tqdm==4.65.0
click==8.1.6
python-dateutil==2.8.2 